enable gcc to check arguments of printf-like functions
Check printf-like functions like conoutf, if format strings and argument
lists match.

This commit also fixes a lot of code parts that actually had mismatching
argument lists - or did things, that this check does not like.

numtime() was rewritten in an ugly but very portable fashion (it's
recommended to stay away from "long long" format entries, because errors
usually only come up at runtime).

(some code was copied from sauer)

+++ source/src/stream.cpp
@@ -611,7 +611,7 @@ struct gzstream : stream
             if(checkcrc != crc)
                 conoutf("gzip crc check failed: read %X, calculated %X", checkcrc, crc);
             if(checksize != zfile.total_out)
-                conoutf("gzip size check failed: read %d, calculated %d", checksize, zfile.total_out);
+                conoutf("gzip size check failed: read %u, calculated %u", checksize, (uint) zfile.total_out);
         }
 #endif
     }
